Redesign song database
This time always record a full URL into each file-entry, which encodes all the details
required to access the file. Two schemes are supported: file and zip. File is regular
file scheme, and zip refers to a song contained in a zip file with query parameter
path used to specify the entry within the zip.

We can now add support to supplying files over the network, provided we write the
http(s) scanner that is capable of indexing relevant network shares, and
define a way to bootstrap the scanning process over the network (LINK files).