Add vgmstream as a plugin.
NOTE: This is semi-working. This needs to be fixed. It will play a song for a certain while (presumably until the buffer fills or some shit like that), and then the song will cut out and simply not play.

I've tried as hard as I can to fix this and can't come up with a solution. I'll keep looking however.