Change some of my comments, and other ones, Mostly with capitalization.
Changed a 'double-comment' in VGMStreamPlugin.java to be a single one.

Also, not sure if this is a correct way to solve an onDraw problem, but I've moved the "Paint fftPaint = new Paint();" out of the method.

This was only done due to Android Lint saying it's bad to do object allocation in a draw method.

If this is a wrong way of solving it, just call me an idiot.